# Conflicts:
#	I am poor and hungry/Library/CurrentLayout.dwlt
#	I am poor and hungry/Library/ScriptAssemblies/Assembly-CSharp-Editor.dll
#	I am poor and hungry/Library/ScriptAssemblies/Assembly-CSharp-Editor.dll.mdb
#	I am poor and hungry/Library/ScriptAssemblies/Assembly-CSharp-firstpass.dll
#	I am poor and hungry/Library/ScriptAssemblies/Assembly-CSharp-firstpass.dll.mdb
#	I am poor and hungry/Library/ScriptAssemblies/Assembly-CSharp.dll
#	I am poor and hungry/Library/ScriptAssemblies/Assembly-CSharp.dll.mdb
#	I am poor and hungry/Library/assetDatabase3
#	I am poor and hungry/Library/expandedItems
#	I am poor and hungry/Library/metadata/00/00000000000000002000000000000000
#	I am poor and hungry/Library/metadata/00/00000000000000004000000000000000
#	I am poor and hungry/Library/metadata/00/00000000000000006100000000000000
#	I am poor and hungry/Library/metadata/22/222631802c1b38149803ddfd137feecb
#	I am poor and hungry/Library/metadata/e5/e56f0fb0dccc002449ab3b787491c95f

+++ I am poor and hungry/Assets/Scripts/Food.cs
@@ -22,7 +22,7 @@ public class Food //: Item
     public int rarity;
     public string[] categories;
     public int food;
-    public int nutrition;
-    public int poison;
+    public float nutrition;
+    public float poison;
     public ModifierHandler modifier;
 }
